Split SdJwtVcVerifierFactory to a separate file
